+++ src/ui/messages/carousel/samples.js
@@ -142,9 +142,9 @@ model.generateMessage($("layer-conversation-view").conversation, message => mess
 
 new TextModel({text: "Product and Button Carousel:"}).send({ conversation: $("layer-conversation-view").conversation });
 
-ProductModel = client.getMessageTypeModelClassForMimeType('application/vnd.layer.product+json')
-ImageModel = client.getMessageTypeModelClassForMimeType('application/vnd.layer.image+json')
-CarouselModel = client.getMessageTypeModelClassForMimeType('application/vnd.layer.carousel+json')
+ProductModel = Layer.client.getMessageTypeModelClassForMimeType('application/vnd.layer.product+json')
+ImageModel = Layer.client.getMessageTypeModelClassForMimeType('application/vnd.layer.image+json')
+CarouselModel = Layer.client.getMessageTypeModelClassForMimeType('application/vnd.layer.carousel+json')
    ButtonsModel = Layer.Core.Client.getMessageTypeModelClass('ButtonsModel')
    ChoiceModel = Layer.Core.Client.getMessageTypeModelClass('ChoiceModel')
 model = new CarouselModel({